move memoized to util
